LayoutTests/fast/dom/HTMLKeygenElement/keygen.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<keygen id="keys" />
<pre id="console">
This tests the keygen element.

</pre>
<script>
var keygen = document.getElementById('keys');
shouldBe('keygen.options', 'undefined');  // keygen is not a select

if (window.internals) {
  var shadow = internals.shadowRoot(keygen);
  shouldBe('shadow.firstChild.tagName', '"SELECT"');
  shouldBe('internals.shadowRoot(shadow.firstChild)', 'null');
}
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
